backend/requirements.txt
# Core
django==5.1.4
django-ninja==1.3.0
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
gunicorn==23.0.0

# Authentication
PyJWT==2.9.0

# Logging
python-json-logger==3.2.1

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Code quality
ruff==0.9.7
